backend/src/requirements.txt
flask
flask-cors
langchain
langchain-core
langgraph
fastapi
uvicorn[standard]
langchain-google-genai
python-dotenv 